moving things to function
