+++ packages/npm/oa42-generator/src/generators/classes/server.ts
@@ -1,15 +1,13 @@
 import * as models from "../../models/index.js";
 import { itt } from "../../utils/iterable-text-template.js";
-import { generateEndpointHandlerMethod, generateRequestHandlerMethod } from "../functions/index.js";
 import {
-  getAuthenticationHandlerName,
-  getAuthenticationHandlerTypeName,
-  getOperationHandlerName,
-  getOperationHandlerTypeName,
-  getRegisterAuthenticationHandlerName,
-  getRegisterOperationHandlerName,
-  getServerAuthenticationTypeName,
-} from "../names/index.js";
+  generateEndpointHandlerMethod,
+  generateRegisterMiddlewareMethod,
+  generateRequestHandlerMethod,
+  registerAuthenticationHandlerMethod,
+  registerOperationHandlerMethod,
+} from "../functions/index.js";
+import { getServerAuthenticationTypeName } from "../names/index.js";
 
 /**
  * Generated the server class. This is the server that is generated from the
@@ -53,82 +51,17 @@ function* generateBody(apiModel: models.Api) {
     }
   `;
 
-  // TODO move to functions
-  yield itt`
-    public registerMiddleware(middleware: lib.ServerMiddleware) {
-      const nextMiddleware = this.middleware;
-  
-      this.middleware =
-        lib.wrapAsync(
-          async (request, next) => await middleware.call(this, request, async (request) =>
-              await nextMiddleware.call(this, request, next)
-            ),
-            this.middlewareWrapper,
-            middleware.name,
-          );
-    }
-  `;
+  yield generateRegisterMiddlewareMethod();
 
   yield generateRequestHandlerMethod(apiModel);
 
   for (const authenticationModel of apiModel.authentication) {
-    const registerHandlerMethodName = getRegisterAuthenticationHandlerName(authenticationModel);
-    const handlerTypeName = getAuthenticationHandlerTypeName(authenticationModel);
-    const handlerPropertyName = getAuthenticationHandlerName(authenticationModel);
-
-    yield itt`
-      private ${handlerPropertyName}?: ${handlerTypeName}<A>;
-    `;
-
-    // TODO add function to register all authentication handlers
-    // TODO move to functions
-    yield itt`
-      public ${registerHandlerMethodName}(authenticationHandler: ${handlerTypeName}<A>) {
-        this.${handlerPropertyName} =
-          lib.wrapAsync(
-            authenticationHandler,
-            this.authenticationWrapper,
-            ${JSON.stringify(authenticationModel.name)},
-          );
-      }
-    `;
+    yield registerAuthenticationHandlerMethod(authenticationModel);
   }
 
   for (const pathModel of apiModel.paths) {
     for (const operationModel of pathModel.operations) {
-      const handlerPropertyName = getOperationHandlerName(operationModel);
-      const handlerTypeName = getOperationHandlerTypeName(operationModel);
-      const registerHandlerMethodName = getRegisterOperationHandlerName(operationModel);
-
-      yield itt`
-        private ${handlerPropertyName}?: ${handlerTypeName}<A>;
-      `;
-
-      // TODO add function to register all operation handlers
-      // TODO move to functions
-      const jsDoc = [
-        operationModel.deprecated ? "@deprecated" : "",
-        operationModel.summary,
-        operationModel.description,
-      ]
-        .map((line) => line.trim())
-        .filter((line) => line.length > 0)
-        .join("\n");
-
-      yield itt`
-        /**
-         ${jsDoc}
-         */
-        public ${registerHandlerMethodName}(operationHandler: ${handlerTypeName}<A>) {
-          this.${handlerPropertyName} =
-            lib.wrapAsync(
-              operationHandler,
-              this.operationWrapper,
-              ${JSON.stringify(operationModel.name)},
-            );
-        }
-      `;
-
+      yield registerOperationHandlerMethod(operationModel);
       yield generateEndpointHandlerMethod(apiModel, operationModel);
     }
   }
