+++ packages/npm/oa42-generator/src/generators/functions/index.ts
@@ -3,4 +3,8 @@ export * from "./endpoint-handler-method.js";
 export * from "./is-authentication.js";
 export * from "./is-request-parameters.js";
 export * from "./is-response-parameters.js";
+export * from "./register-authentication-handler.js";
+export * from "./register-middleware.js";
+export * from "./register-operation-handler.js";
 export * from "./request-handler-method.js";
+
