+++ packages/npm/oa42-generator/src/generators/functions/register-middleware.ts
@@ -1,0 +1,19 @@
+import { itt } from "../../utils/index.js";
+
+export function *generateRegisterMiddlewareMethod(){
+  yield itt`
+  public registerMiddleware(middleware: lib.ServerMiddleware) {
+    const nextMiddleware = this.middleware;
+
+    this.middleware =
+      lib.wrapAsync(
+        async (request, next) => await middleware.call(this, request, async (request) =>
+            await nextMiddleware.call(this, request, next)
+          ),
+          this.middlewareWrapper,
+          middleware.name,
+        );
+  }
+`;
+
+}
